MacOSX: (g2o) Fixed cholmod library not found. New ICP parameter: "Icp/Epsilon". Fixed null ICP_inliers_ratio with odometry statistics.

+++ corelib/include/rtabmap/core/OdometryInfo.h
@@ -53,7 +53,7 @@ public:
 	bool lost;
 	int matches;
 	int inliers;
-	int icpInliersRatio;
+	float icpInliersRatio;
 	float variance;
 	int features;
 	int localMapSize;
